requirements.txt
Flask==3.1.1
psycopg2==2.9.10
PyYAML==6.0.2
asgiref==3.8.1
dotenv==0.9.9
uvicorn==0.34.2
